requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.2.2
faiss-gpu==1.7.2
numpy==1.24.3
pillow==10.2.0
tqdm==4.66.1
torch==2.1.0
